% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_from_raw.R
\name{index_from_raw}
\alias{index_from_raw}
\title{Calculates index scores from raw domain data input. Assumes first column is id}
\usage{
index_from_raw(
  ds,
  indx = cognitive.index.lookup::index_table,
  version.col = "ab",
  age.col = "age",
  raw_columns = c("imm", "vis", "ver", "att", "del"),
  mani = TRUE
)
}
\arguments{
\item{ds}{data set}

\item{indx}{index file}

\item{version.col}{version column name. The function will check formatting}

\item{age.col}{age column name}

\item{raw_columns}{column names of raw data columns}

\item{mani}{flag to manipulate index data for plotting}
}
\value{
data.frame
}
\description{
Calculates index scores from raw domain data input. Assumes first column is id
}
\examples{
ds <- cognitive.index.lookup::sample_data
ds |> index_from_raw()
read_input("https://raw.githubusercontent.com/agdamsbo/cognitive.index.lookup/main/data/sample.csv") |> index_from_raw()
}
