% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{percentile_sub_plot}
\alias{percentile_sub_plot}
\title{Plots percentile}
\usage{
percentile_sub_plot(data, grp.color)
}
\arguments{
\item{data}{data set}

\item{grp.color}{column to color}
}
\value{
ggplot2 list object
}
\description{
Plots percentile
}
