% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index2}
\alias{plot_index2}
\title{Plot index and percentile}
\usage{
plot_index2(data, ...)
}
\arguments{
\item{data}{data set}

\item{...}{arguments passed}
}
\value{
ggplot list
}
\description{
Plot index and percentile
}
\examples{
sample_data |> 
 index_from_raw() |> 
 dplyr::filter(id \%in\% 1:5) |>
 dplyr::filter(ab==1) |>
 plot_index2(facet.by="ab")

}
