% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_prep}
\alias{plot_prep}
\title{Data preparation for index score plotting}
\usage{
plot_prep(data, sub_plot, scores, grp.color, facet.by, dom_names)
}
\arguments{
\item{data}{data set}

\item{sub_plot}{sub plot ("_is" or "_per")}

\item{scores}{score variable name bits}

\item{grp.color}{column to color}

\item{facet.by}{variable to facet by}

\item{dom_names}{domain names}
}
\value{
data frame
}
\description{
Data preparation for index score plotting
}
