% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_stitch}
\alias{plot_stitch}
\title{Stitch ggplots}
\usage{
plot_stitch(list, fun = index_sub_plot, grp.color)
}
\arguments{
\item{list}{list of plots}

\item{fun}{function to use for plotting}

\item{grp.color}{column to color}
}
\value{
ggplot2 list object
}
\description{
Stitch ggplots
}
