% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_index.R
\name{shiny_index}
\alias{shiny_index}
\title{Launch the included Shiny-app for index calculations}
\usage{
shiny_index()
}
\value{
shiny app
}
\description{
Launch the included Shiny-app for index calculations
}
